Sprint 12. Fix reviewer notes.
